NotFittedError now only inherits from sklearn
It can be confusing if a trace contains a
"sklearn.exceptions.NotFittedError".
Users may search for the problem in sklearn,
when we really want to indicate a misusage
of *our* estimator code. To alleviate such
problems while still profiting from previous
design of this error class we simply inherit
from `sklearn.exceptions.NotFittedError` but
use `causality.exceptions.NotFittedError` to
indicate problems with unfitted causality estimators.